[Feat] Home Page 이미지 및 현황정보  SWR 적용
