+++ src/shared/hook/useFittyText.ts
@@ -20,5 +20,5 @@ export default function useFitty(
         fittyInstance.unsubscribe();
       };
     }
-  }, []);
+  }, [titleRef]);
 }
